fixing module structure for utils
